Move transient props to types folder
