+++ desktop/packages/mullvad-vpn/src/renderer/lib/types/index.ts
@@ -1,0 +1,1 @@
+export * from './transient-props';
